+++ Singe/Services/old/CommandAttribute.cs
@@ -2,7 +2,7 @@
 using System.Collections.Generic;
 using System.Text;
 
-namespace Singe.Services
+namespace Singe.Services.old
 {
     public sealed class CommandAttribute : Attribute
     {
