+++ Singe/Services/old/CommandSignature.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Reflection;
 using System.Text;
 
-namespace Singe.Services
+namespace Singe.Services.old
 {
     class CommandSignature
     {
