+++ Singe/Services/old/SingeService.cs
@@ -2,7 +2,7 @@
 using System.Collections.Generic;
 using System.Text;
 
-namespace Singe.Services
+namespace Singe.Services.old
 {
     public class SingeService : Service
     {
